Created a functionality to find the closest hospital to the user according to the database.
Added functional markers for the user and the hospitals. The markers have the name and wait time of the hospital (morning only for prototype) and they show which of the hospital markers is the closest to the user.

Fixed link so when the user presses hospital wait time in the map.html page, they are sent to the page waittime.html

Added a reviewLoc.html for when the user gets closer to the hospital's location (no javascript or addition to the database has been properly implemented)